% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_ceiling.R
\name{get_numeric_transform}
\alias{get_numeric_transform}
\title{Get appropriate numeric transformation function}
\usage{
get_numeric_transform(transform, ...)
}
\arguments{
\item{transform}{one of the following formats:
\itemize{
\item \code{character} string matching a recognized transformation:
\itemize{
\item \code{"log2signed"}
\item \code{"exp2signed"}
\item \code{"sqrt"}
\item \code{"square"}
\item \code{"cubert"}
\item \code{"cube"}
\item \code{"qrt"}, \code{"quadrt"}
\item \code{"frt"}, \code{"fthrt"}
\item recognized mathematical function names (in scope), such as
\code{log()}, \code{log10()}, etc. Any function retrieved by \code{get()}
can be used, provided it returns data in the same dimensions
as provided.
}
\item \code{function}: this function is applied to the \code{numeric} matrix
with no additional modifications. It is expected to return
a \code{numeric} matrix with exact same dimensions.
\item \code{list} that contains a series of values with either
\code{character} or \code{function} as described above.
}}

\item{...}{additional arguments are ignored.}
}
\value{
\code{function} or \code{NULL} when no matching function is
found, or \code{list} is returned when the input \code{transform}
has multiple values.
}
\description{
Get appropriate numeric transformation function
}
\details{
This function recognizes numeric transformation functions by
name, or searches the attached R package environments for
a matching function by name.

Recognized transform names:
\itemize{
\item \code{"none"} or \code{"linear"} returns the data without change
\item \code{"log2signed"} applies \code{jamba::log2signed()}, defined as
\code{log2(1+x)} transform to the absolute value, then multiplies
by the original \code{sign(x)}
\item \code{"exp2signed"} applies the inverse of \code{"log2signed"}, which
exponentiates numeric values which were previously transformed
with \code{log2(1+x)}. Note that value=0 when exponentiated becomes \code{+1}
\item \code{"sqrt"} applies square root transform, equivalent to \code{sqrt()}
\item \code{"cubert"} applies cube root \code{x^(1/3)}
\item \code{"qrt"} applies fourth root \code{x^(1/4)}
\item \code{"frt"} or \code{"fthrt"} applies fifth root \code{x^(1/5)}
\item \code{"square"} applies \code{x^2} to absolute value, multiplied by the \code{sign(x)}
\item \code{"cube"} applies \code{x^3};
}

Any other character name is used to find a function with the same
name, and if the function is found it is used. The function is
expected to take input \code{x} and return corresponding output with the
same length as the input. A function such as \code{max()} does not fit
this criteria, but a function such as \code{log2()} is acceptable.
}
\examples{
get_numeric_transform("log2signed")

transform_list <- get_numeric_transform(
   c("none",
   "log2signed",
   log2,
   sqrt));
jamba::sdim(transform_list);

x <- jamba::nameVector(0:16);
transform_list[[1]](x)

transform_list[[2]](x)

}
\seealso{
Other jam utility functions: 
\code{\link{color_complement}()}
}
\concept{jam utility functions}
